refactor: extract golden-corpus time-shift logic into timeshift.go
No behavior change (verified: 2 more full golden-corpus runs, 776/776
green, plus go test ./... clean) — just moves referenceCaptureTime,
captureWindow, the DateTime layouts, shiftableFields, shiftTimeString,
and shiftRowTimestamps out of main.go into their own file, and wraps the
shift computation in a small captureShift() helper.

The golden-corpus/parity harness is expected to be retired gradually as
the Go server diverges from byte-parity with the frozen Python oracle —
new/changed behavior going forward should get conventional Go tests
(storetest.FakeDB unit tests, semantic integration tests) instead of more
golden profiles. Until that migration finishes, this file is the one
place the capture-session time-shifting logic lives, so it reads as a
deliberately maintained test utility rather than being buried inline in
the seeder's main().
